# apigateway/docs/swagger.yaml
definitions:
...
  requests.CreateOrderItemRequest:
    properties:
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - price
    - product_id
    - quantity
    type: object
  requests.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/requests.CreateOrderItemRequest'
        type: array
      total_price:
        type: integer
      user_id:
        type: integer
    required:
    - items
    - total_price
    - user_id
    type: object
...
  requests.UpdateOrderItemRequest:
    properties:
      order_item_id:
        type: integer
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - order_item_id
    - price
    - product_id
    - quantity
    type: object
  requests.UpdateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/requests.UpdateOrderItemRequest'
        type: array
      order_id:
        type: integer
      total_price:
        type: integer
      user_id:
        type: integer
    required:
    - items
    - total_price
    - user_id
    type: object
...
  response.ApiResponseOrder:
    properties:
      data:
        $ref: '#/definitions/response.OrderResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseOrderAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseOrderDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseOrderDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.OrderResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponsePaginationOrder:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationOrderDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ApiResponsesOrderItem:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderItemResponse'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  response.OrderItemResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.OrderResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/order:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrder'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all orders
      tags:
      - Order
...
  /api/order-item/order/{order_id}:
    get:
      consumes:
      - application/json
      description: Retrieve order items by order ID
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of order items by order ID
          schema:
            $ref: '#/definitions/response.ApiResponsesOrderItem'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve order item data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find order items by order ID
      tags:
      - OrderItem
...
  /api/order/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order data
          schema:
            $ref: '#/definitions/response.ApiResponseOrder'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find order by ID
      tags:
      - Order
  /api/order/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active orders
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrderDeleteAt'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active orders
      tags:
      - Order
  /api/order/create:
    post:
      consumes:
      - application/json
      description: Create a new order with provided details
      parameters:
      - description: Order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created order
          schema:
            $ref: '#/definitions/response.ApiResponseOrder'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create order
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new order
      tags:
      - Order
  /api/order/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete an order record by its ID.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted order record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseOrderDelete'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete order:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete an order
      tags:
      - Order
  /api/order/delete/all:
    post:
      consumes:
      - application/json
      description: Permanently delete all order records.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted all orders permanently
          schema:
            $ref: '#/definitions/response.ApiResponseOrderAll'
        "500":
          description: Failed to delete orders
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete all orders
      tags:
      - Order
  /api/order/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed order record by its ID.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored order
          schema:
            $ref: '#/definitions/response.ApiResponseOrderDeleteAt'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore order
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed order
      tags:
      - Order
  /api/order/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all trashed order records.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all orders
          schema:
            $ref: '#/definitions/response.ApiResponseOrderAll'
        "500":
          description: Failed to restore orders
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all trashed orders
      tags:
      - Order
  /api/order/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed orders
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrderDeleteAt'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed orders
      tags:
      - Order
  /api/order/trashed/{id}:
    post:
      consumes:
      - application/json
      description: Retrieve a trashed order record by its ID.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trashed order
          schema:
            $ref: '#/definitions/response.ApiResponseOrderDeleteAt'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve trashed order
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve a trashed order
      tags:
      - Order
  /api/order/update:
    put:
      consumes:
      - application/json
      description: Update an existing order with provided details
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated order
          schema:
            $ref: '#/definitions/response.ApiResponseOrder'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update order
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing order
      tags:
      - Order
...
swagger: "2.0"